fix(common): CHECKOUT-3334 Prevent order data from overwriting checkout data when retrieving payment information
When we load an order to check if we have to finalize it, firing the
reducers causes billing and coupons data to be overwritten by the order
information.

This commit adds a new action type `LoadOrderPayments` that allows for
payment information to be loaded without causing all reducers to fire.