refactor(payment): CHECKOUT-3331 Mark `CheckoutService#loadPaymentMethod` as internal
BREAKING CHANGE: `loadPaymentMethod` has been deprecated for some time.
However, due to legacy reasons, this method is still in use by Optimized
Checkout. To further discourage people from using it, we now mark it as
`internal` and remove it from the documentation. We will completely
remove it once it is no longer used by Optimized Checkout.