perf(common): CHECKOUT-3462 Apply `cacheAction` decorator to certain actions
Previously, if an action is executed twice in parallel, it won't be able
to return cached data because the cache is not available at the time of
execution. Now, with this change, it can because the decorator
internally caches the action and replays the cached value for the second
call as soon as it becomes available.