fix(checkout): INT-5409 Change error message when payment data is unavailable
When trying to modify a cart for an order already payed an error message
is shown informing the user that payment method is not available.

This modify the message in order to be more verbose with the customer,
as the error occurs due to missing cart information or if the order has
already been paid.